memcpy

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: yolee <yolee@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/24 17:38:55 by yolee             #+#    #+#             */
-/*   Updated: 2021/11/26 20:07:16 by yolee            ###   ########.fr       */
+/*   Updated: 2021/12/16 16:44:48 by yolee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int				loop;
+	size_t			loop;
 	unsigned char	*conv_s;
 
 	conv_s = (unsigned char *)s;
