+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: yolee <yolee@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/24 17:39:34 by yolee             #+#    #+#             */
-/*   Updated: 2021/11/26 16:24:53 by yolee            ###   ########.fr       */
+/*   Updated: 2021/12/16 16:57:37 by yolee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int				loop;
+	size_t			loop;
 	unsigned char	conv_c;
 	unsigned char	*conv_b;
